restful api version done

+++ src/main/java/fun/kevins/todos/service/TodoService.java
@@ -5,9 +5,7 @@ import fun.kevins.todos.entity.Todo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.text.SimpleDateFormat;
-import java.util.Date;
-import java.util.TimeZone;
+import java.util.Optional;
 
 @Service
 public class TodoService {
@@ -18,39 +16,43 @@ public class TodoService {
         return todoDao.findAll();
     }
 
-    public Iterable<Todo> createTodo(Todo todo) {
-        SimpleDateFormat df = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
-        df.setTimeZone(TimeZone.getTimeZone("GMT"));
-        String date = df.format(new Date());
-        todo.setCreateTime(date);
-        todo.setUpdateTime(date);
-        todoDao.save(todo);
-        return getTodos();
+    public Integer createTodo(Todo todo) {
+        Todo rltTodo = todoDao.save(todo);
+        return rltTodo.getId();
     }
 
-    public Todo updateTodo(Integer id, Todo todo) {
-        try {
-            Todo resTodo = findById(id);
-            Integer status = todo.getStatus();
-            resTodo.setStatus(status);
-            return todoDao.save(resTodo);
-        } catch (Exception exception) {
-            return null;
+    public Boolean updateTodo(Integer id, Todo todo) {
+        Optional<Todo> isExistTodo = findById(id);
+        if (!isExistTodo.isPresent()) {
+            return false;
         }
-
+        Todo newTodo = isExistTodo.get();
+        if (todo.getStatus() == null) {
+            return false;
+        }
+        newTodo.setStatus(todo.getStatus());
+        try {
+            todoDao.save(newTodo);
+            return true;
+        } catch (Exception e) {
+            return false;
+        }
     }
 
-    public Todo findById(Integer id) {
-        Todo todo = todoDao.findById(id).get();
+    public Optional<Todo> findById(Integer id) {
+        Optional<Todo> todo = todoDao.findById(id);
         return todo;
     }
 
     public Boolean deleteTodo(Integer id) {
+        Optional<Todo> findTodo = findById(id);
+        if (!findTodo.isPresent()) {
+            return false;
+        }
         try {
-            Todo resTodo = findById(id);
             todoDao.deleteById(id);
             return true;
-        } catch (Exception exception) {
+        } catch (Exception e) {
             return false;
         }
     }
